Move token related code to the polaris/token package (#63)
