+++ internal/errors/json_error.go
@@ -1,0 +1,41 @@
+// Copyright 2021 Rubrik, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to
+// deal in the Software without restriction, including without limitation the
+// rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+// sell copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+// DEALINGS IN THE SOFTWARE.
+
+package errors
+
+import "fmt"
+
+// JSONError is returned by Polaris in the body of a response as a JSON
+// document when certain types of errors occur.
+type JSONError struct {
+	Code    int    `json:"code"`
+	URI     string `json:"uri"`
+	Message string `json:"message"`
+}
+
+// IsError determines if the JSONError unmarshalled from a JSON document
+// represents an error or not.
+func (e JSONError) IsError() bool {
+	return e.Code != 0 || e.Message != ""
+}
+
+func (e JSONError) Error() string {
+	return fmt.Sprintf("%s (code %d)", e.Message, e.Code)
+}
