+++ internal/errors/testdata/error_json_from_polaris.json
