+++ pkg/polaris/graphql/aws/testdata/validate_and_create_aws_cloud_account.json
