+++ pkg/polaris/token/local_user_source.go
@@ -18,7 +18,7 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-package graphql
+package token
 
 import (
 	"encoding/json"
@@ -29,63 +29,50 @@ import (
 	"github.com/rubrikinc/rubrik-polaris-sdk-for-go/pkg/polaris/log"
 )
 
-// serviceAccountSource holds all the information needed to obtain a token
-// for a service account.
-type serviceAccountSource struct {
-	logger       log.Logger
-	client       *http.Client
-	tokenURL     string
-	clientID     string
-	clientSecret string
+// LocalUserSource holds all the information needed to obtain a token for a
+// local user account.
+type LocalUserSource struct {
+	logger   log.Logger
+	client   *http.Client
+	tokenURL string
+	username string
+	password string
 }
 
-// newServiceAccountSource returns a new token source that uses the
-// http.DefaultClient to obtain tokens.
-func newServiceAccountSource(accessTokenURL, clientID, clientSecret string, logger log.Logger) *serviceAccountSource {
-	return &serviceAccountSource{
-		logger:       logger,
-		client:       http.DefaultClient,
-		tokenURL:     accessTokenURL,
-		clientID:     clientID,
-		clientSecret: clientSecret,
+// NewLocalUserSource returns a new token source that uses the specified client
+// to obtain tokens.
+func NewLocalUserSource(client *http.Client, apiURL, username, password string, logger log.Logger) *LocalUserSource {
+	return &LocalUserSource{
+		logger:   logger,
+		client:   client,
+		tokenURL: fmt.Sprintf("%s/session", apiURL),
+		username: username,
+		password: password,
 	}
 }
 
-// token returns a new token from the service account token source.
-func (src *serviceAccountSource) token() (token, error) {
+// token returns a new token from the local user token source.
+func (src *LocalUserSource) token() (token, error) {
 	// Prepare the token request body.
 	body, err := json.Marshal(struct {
-		GrantType    string `json:"grant_type"`
-		ClientID     string `json:"client_id"`
-		ClientSecret string `json:"client_secret"`
-	}{GrantType: "client_credentials", ClientID: src.clientID, ClientSecret: src.clientSecret})
+		Username string `json:"username"`
+		Password string `json:"password"`
+	}{Username: src.username, Password: src.password})
 	if err != nil {
 		return token{}, fmt.Errorf("failed to marshal token request body: %v", err)
 	}
 
-	var resp []byte
-	for attempt := 1; ; attempt++ {
-		src.logger.Printf(log.Debug, "Acquire access token (attempt: %d)", attempt)
-		resp, err = requestToken(src.client, src.tokenURL, body)
-		if err == nil {
-			break
-		}
-		if !errors.Is(err, errTokenRequestTimeout) || attempt == tokenRequestAttempts {
-			return token{}, fmt.Errorf("failed to acquire service account access token: %v", err)
-		}
+	resp, err := Request(src.client, src.tokenURL, body, src.logger)
+	if err != nil {
+		return token{}, fmt.Errorf("failed to acquire local user access token: %v", err)
 	}
 
-	// Try to parse the JSON document as an access token. Verify that the
-	// response has the same client id as the request.
+	// Try to parse the JSON document as an access token.
 	var payload struct {
-		ClientID    string `json:"client_id"`
 		AccessToken string `json:"access_token"`
 	}
 	if err := json.Unmarshal(resp, &payload); err != nil {
 		return token{}, fmt.Errorf("failed to unmarshal token response body: %v", err)
-	}
-	if payload.ClientID != src.clientID {
-		return token{}, errors.New("invalid client id")
 	}
 	if payload.AccessToken == "" {
 		return token{}, errors.New("invalid token")
